WIP: Installing boost-spirit "properly"
Boost-spirit is not a self-contained library by any means,
unfortunately. Rather than inventing ad-hoc measures of installing the
dependencies and dependencies of dependencies and..., let's stop playing
whack-a-mole and instead use the big installation hammer.

That CMake ugliness is there because the FindBoost.cmake doesn't appear
to support finding header-only Boost with no libraries.

FIXME: remove that #include

Change-Id: Iad867386ac5907630db8bdde0668d8fd9edee580