WIP: Installing boost-spirit "properly"
Boost-spirit is not a self-contained library by any means,
unfortunately. Rather than inventing ad-hoc measures of installing the
dependencies and dependencies of dependencies and..., let's stop playing
whack-a-mole and instead use the big installation hammer.

We're not calling `b2 install` because that one attempts to build these
libraries. We do not want that. On the other hand, `b2 headers` is
"lazy" as it prefers to create symlinks where it makes sense. These
symlinks are relative, pointing to the source directory of each
individual Boost module -- hence that `cp`.

That CMake ugliness is there because the FindBoost.cmake doesn't appear
to support finding header-only Boost with no libraries.

FIXME: remove that #include

Change-Id: Iad867386ac5907630db8bdde0668d8fd9edee580