Added -Wsuggest-override and -Woverloaded-virtual warnings
We will adhere to the convention of marking all overriden methods
with the "override" specifier, to make clear which methods
are virtual overriden and which are not. -Wsuggest-override
emits warning any time the compiler finds a method which
should be marked "override" but is not.

We also add the -Woverloaded-virtual warning to inform
us when we try to hide a base class' virtual method,
because that will be almost certainly a bug rather
than a genius hack.

Change-Id: I19369689ce89e99b5abce1a0cc8b6ce68886fe41