Change how path completion suggests nodes
Path completions now include module names only if the top-level module
is different. Module names in the middle of paths are redundant and it
makes auto-completed paths shorter. Ideally, the program should be able
to complete nodes without module prefixes and, optionally, if the user
started typing a module name, be able to complete nodes with module
prefixes. However, I don't think it is such an important feature,
because, as was said, module names are redundant (if equal to top-level
module), so the completion doesn't suggest it.

Change-Id: Iedabd017d0415a919a45be4aca30a5b353193f28