Fix ParserContext flag not properly resetting
The m_topLevelModulePresent was not being properly reset when the
starting slash of an absolute path was parsed. For example:
/ietf-interfaces:interfaces> cd /i
failed when trying to parse the "i".
The parser thought that a top-level module was present and while in
fact, "/i" doesn't have a top-level module present. The flag is now
properly beign reset when parsing the startign slash of an absolute
path.

https: //tree.taiga.io/project/jktjkt-netconf-cli/issue/99
Change-Id: If5144cffcf55acbf869cf24a907d2c8c174724bd