Change how create/delete path argument is parsed
The handler for create/delete was an inherited class from
presenceContainerPathHandler class, so that the code is shared between
create/delete. However, this means that it is impossible to change the
grammar for create/delete to include other types of paths (because the
handler was hardcoded to presence containers). This change ties the
handler to the "path" rule rather than to the create/delete rule. This
method is also in line with the "set" command where the path is checked
in the "leafPath" rule rather than in the "set" rule.

Change-Id: Ic41aa26bf4bc93f762435f43312489e10472e475