Implement int8, uint8, int16, uint16, int64, uin64 types
This sw did not support uint8 type, which is needed by the upcoming
driver CalibrationBox within cla-sysrepo, so we implemented uint8 and
other types in TODO sections.

It is unclear to us whether it is really necessary to add all the
leaf_data_XXX_class classes -- feel free to propose something better,
Vasek.

Change-Id: I866a3933fe21ea7844299556e5aaf39b3e40e92f
Co-Authored-By: Jan Kundrát <jan.kundrat@cesnet.cz>