Improve ctest logic
Firstly, as there will be more datastore tests, sysrepo.cpp had to be
changed to something different.

The overall logic for the datastore tests had to be changed. Now, all
datastore tests depend on the start_daemons (using fixtures). This means
that even if one starts the `test_setting_values_sysrepo` test, it
always runs start_daemons (the previous config only ensured, that
start_daemons ran before the datastore tests).

The actual datastore test creation was also refactored
to a function. This ensures that both sysrepo and netconf versions of
the test will get created.

Change-Id: Id74c9671802ed746d44f33ba0026eb81a1594e2a