Merge pathToAbsoluteSchemaString and pathToSchemaString
This functions did almost the same thing, the only difference was that
the former inserted module prefixes everyhwere and the other didn't.
The new function now takes an argument specifying whether to insert
prefixes everyhwere.

The fact that this was split into two function led to an error in a
later patch.

Change-Id: I6112cb5982919b51b61152f355771a52dd467c6e