Filter by path in NetconfAccess::getItems
libnetconf might not always return just the data nodes specified by
getItems' argument. One such example is when getting a leaf which is key
inside a list: all the other will be present. To avoid this, I apply the
`path` again. This is not the case with the sysrepo library, because it
does this internally too.

Change-Id: I27b5df9afa0c47897d283bc71f0c31c48fd1b847