Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

We can merge this if you want to (and it's indeed an improvement that it unifies behavior of SR and NC backends). In the long run, this method will need a two-mode flag ("return everything including this and all that is below", which is the current behavior, and "just get me this requested thing and nothing else" that is gonna be used for key completion).

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@e32f8df8-3db0-4f76-a6c6-fe9d6e69dd68>
Label: Code-Review=+2