Update patch set 1
Patch Set 1: Code-Review-2

If I remember correctly, the final outcome of a face-to-face discussion is that this is not needed anymore now that listInstances() is implemented via a facade on top of DatastoreAccess. This will result in different behavior (visible to the user) when requesting stuff such as:

 get /example:list[name='jane'][surname='smith']/name

where each backend returns different amount of data (libnetconf2 will include /example:list[name='jane'][surname='smith']/surname as well). This is OK for us because we think this is an edge case, and the long-term plan is to eventually migrate to new sysrepo, anyway, which provides us with libyang data structures. That should make the difference go away.

Please restore this and let me know if I remember incorrectly.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@e32f8df8-3db0-4f76-a6c6-fe9d6e69dd68>
Label: Code-Review=-2