Update patch set 49
Patch Set 49:

> Understood. But it's weird.

Well, it found a real problem in this patch -- if it got merged as-is, then master would not pass tests.

This checking is "correct" as in "if it rejects, then it's broken". It does not guard against the target branch moving after the initial Verified+1 vote, so chances are that it will let a "broken" change pass and get merged, based on patch flow timing. That one is covered by https://zuul-ci.org/docs/zuul/discussion/gating.html , but we currently do not use that because of ENOTIME.

Patch-set: 49
Reviewer: Gerrit User 1000000 <1000000@e32f8df8-3db0-4f76-a6c6-fe9d6e69dd68>