libnetconf C++: add support for user RPCs
The API is rather straightforward, it takes a string. I experimented
with consuming a libyang data node instead, but this kept failing
because I was naively creating nodes from a different libyang context,
and that always resulted in an error.

Now, given that the edit-config method already takes a string, let's
make our lives easier and work on this via strings.

Change-Id: I35467dece535df41e223013e32500596d272560c