Make cli behave more like bash
Replxx has the CTRL-W and ALT-Backspace keybinds swapped to do this
opposite from bash. I think the cli should do same thing bash does.
Also, bash recognizes word-splitting characters with the `isalphanum`
function. I can't think of a good way to replicate this with replxx, so
I set the word-splitting characters to a set a use as special characters
in the cli's syntax.

Change-Id: Ibecb71a541112879ab3a6ae0009c3da39632f259