Update patch set 2
Patch Set 2:

Re our discussion about where to inherit the prefixes from:

- https://tools.ietf.org/html/rfc7951#section-6.11 uses the following example:

   /ietf-interfaces:interfaces/interface[name='eth0']/ietf-ip:ipv4/ip

(note the lack of prefix for the last item, the ietf-ip:ip)

- As for what libyang is doing, it looks like the paths are different for schema and data paths :) for increased fun, see https://netopeer.liberouter.org/doc/libyang/master/howtoxpath.html

If you ask me, this is a mess that cannot be directly used, as-is, in a user-facing CLI. It just makes no sense to have a *REQUIRED* difference between:

 /> ls /ietf-interfaces:interfaces/interface/ietf-ip:ipv4/ietf-ip:ip
                                                          ^^^^^^^^^^

while, on the other hand, insisting on:

 /> get /ietf-interfaces:interfaces/interface[name='eth0']/ietf-ip:ipv4/ip
                                                                        ^^

I think that it's reasonable to allow the user to use the prefixes everywhere (including the list keys, i.e.:

 /> get /ietf-interfaces:interfaces/interface[ietf-interfaces:name='eth0']/ietf-ip:ipv4/ip
                                              ^^^^^^^^^^^^^^^^^^^^

On a question of "how to determine a prefix of a schema node when no namespace is given for it and there are at least two preceding namespaces", such as this data path:

 /ietf-interfaces:interfaces/interface/ietf-ip:ipv4/ip
  ^^^^^^^^^^^^^^^                      ^^^^^^^
        what's the prefix of this? -----------------^^

I have a *feeling* that people want to type less, not more, and that most users of this CLI will use it for manipulating data as opposed to querying schema. Also, I cannot imagine a situation where one could "inject" an alien node into a path, i.e., something like this:

 /a:example/a:people[a:name='xyz']/a:address/b:city/a:geolocation
                                             ^^^^^^

So I *think* that there should be no ambiguity when presented with a path like:

 /a:example/people[name='xyz']/address/b:city/geolocation

Clearly, "example", "people" and "address" belong to module "a", and "city" and "geolocation" are from module "b". It does not matter if the "b" module perhaps uses a `typedef` such as `a:geolocation` for its definition of "geolocation" that is nisted to "b:city" -- I think that one is covered by https://tools.ietf.org/html/rfc7950#section-5.4 .

Either way, this is very important, but I think it's orthogonal to tab completion of list keys right now. Let's ignore submodules and prefix differences until the tab completion is in.

Patch-set: 2
Reviewer: Gerrit User 1000000 <1000000@e32f8df8-3db0-4f76-a6c6-fe9d6e69dd68>