Fix path completion on schema paths
The problem was that the grammar was deisgned wrongly. Before this, we
allowed the parser to leave some part of the string unparsed, if
completion was active. However, that lead to some problems where the
stopped prematurely on data path (because we allowed it to end with
unparsed output) while it was possible, to parse more stuff as a schema
path (for example a list without its keys).

The new grammar is a better take on this: first of all
createPathSuggestions is only used if there is a trailing slash (if
there isn't, the completions get created inside the dataNode parser).
Then, if we are doing completion, we parse characters until a next slash
or space. If there actually is another slash, it means that this type of
path wasn't parsed completely, so this whole path can be possibly parsed
even more with other types of path parsers.

Issue: https://tree.taiga.io/project/jktjkt-netconf-cli/issue/151
Change-Id: I7eb1712b28d987dde1192c8438dd055bed71360a