cmake: Fix Boost dependencies
ast_values uses Boost in its public header file, so this dependency has
to be propagated to all of its users. The situation is similar with
ast_path, but it's already getting this public dep via ast_values, but
given that its public headers themselves include Boost directly, let's
behave nicely and make this public dependency explicit.

On the other hand, the utils library does not use Boost directly, so it
can just rely on that transitive dependency from ast_values.

Change-Id: I0c088aa27359d4d972214069a70e27aad99ee7b5