Update patch set 9
Patch Set 9: Code-Review-1

> Patch Set 9:
> 
> > Patch Set 6:
> > 
> > I have no idea, what's causing the compiler error (it works on my machine), but it has something to do with '#include "ast_values.hpp"' in leaf_data_type.hpp
> 
> I really have no idea, it's not the include ast_values.hpp, it's something about how linking in cmake works (I don't know how it works >:( )

In CMake, some library dependencies are not transitive. The CI was failing because you were not getting Boost when building all TUs that include ast_values.hpp, so the proper fix is to do:

 target_link_libraries(ast_values PUBLIC Boost::boost)

This ensures that anything which includes ast_values.hpp will have Boost enabled. I think that you were not hitting this in your dev env because you Boost is probably a systemwide installation in a standard include path such as /usr/include, and therefore it's always enabled "by accident". CI puts Boost into a non-standard path because we want to pin it to a specific version, and that's why it was failing for CI builds. I've fixed this (and some more) in an extra commit, and I'll be rebasing this one on top of that fix.

There were additional issues, too. YangSchema is using stuff from leaf_data_type.hpp, such as Enum and IdentityRef. I do not know where you would like to put these structs, so I created a standalone library for these. I'm OK with putting it into the Schema lib, but I'll leave this to you to decide.

Patch-set: 9
Reviewer: Gerrit User 1000000 <1000000@e32f8df8-3db0-4f76-a6c6-fe9d6e69dd68>
Label: Code-Review=-1