Zuul: Fix artifact retrieval
With our updated Zuul (and the associated update of the base job), a
Zuul manifest is almost always returned as an artifact form all builds.
This has implications with out dependency caching -- the code previously
assumed that we could just pick *the* artifact from the most recent
change item from the CzechLight/Dependencies repo, and that would be our
tarball. That's no longer the case, it could as well be the Zuul
manifest, so let's bring in more JSON magic:

- `.artifacts[]?` is a generator which return the "artifacts" member of
the input as individual items in a sequence, or a null on error ("no
artifacts")
- `select(.foo == "bar")` filters by attribute value, that one is easy,
- `[.artifacts[]? | select ... ]` converts that generator back to an array,
- `[-1]?` selects the last item, if there's any, otherwise returning null,
- ...which is then turned into an empty string, or the resulting value
by `+ ""`.

Change-Id: Ieafa1be1f7066e5005a8e35f4fe86887cbaff3e5