Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The problem here is that I lose the tests for Parser::availableNodes (that was a public method), because now the implementation is inside Interpreter and while Intepreter does have operator(const ls_&) method, it doesn't return anything (it only prints to stdout), so no way to test unless I refactor this somehow.

Solution 1:
I can move the method to Interpreter but it doesn't make too much sense to have a public method, that would be used only for tests (and then only call that method inside operator()(ls_)). Also Interpreter is a visitor, so it makes sense for it to only have operator() methods.

Solution 2:
Change return value of operator()(ls_) to the strings to be printed and print that in main. The operator could then be tested. However, there would have to be some additional if statements and possibly there would have to be a new Visitor class, because of the return value.

I don't like any of these solutions... the first one makes more sense though.

Patch-set: 1