CI: prevent fast cancellation of other jobs upon the first failure
Due to the way how Zuul works with artifacts, any failure during the
whole "job stack" (i.e., all preceding commits) will cause a failure of
the coverage diff.

That would not hurt that much if our tests were completely non-flaky. If
that was the case, the tests would only fail on real errors, and any fix
would require rebasing of follow-up commits anyway. In our reality,
however, we're forced to do a recheck every now and then, and when that
happens, the subsequent changes do not get any update in their coverage
report.

Revert that early-exit, and as a result, a failure in, let's say, ASAN
or TSAN builds that is only transient will not cause a failure of the
coverage diffing in subsequent commits.

Change-Id: I6eba9025aab57f5b462aa1c33e86096a551d0aea