Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > Patch Set 11:
> > 
> > Yikes. Huh. Well, it's probably because I didn't add enough types to YangSchema::nodeType. I should've probably just added all of them. It could be anydata, but idk. I would have to know the type of the node nodeType is being called with. :/
> 
> It's actually getting a nullptr, i.e., the shared_ptr is a null pointer.

Are you sure? I can see in the backtrace, that the throw is happening in the default case of the switch statement for libyang types, which means YangSchema got a type it cannot handle. If it were a nullptr, the throw would happen on line 410. I must admit that having my custom exception there gives me nothing and also is actively confusing, since InvalidNodeException also means "I can't handle this YANG type".

Patch-set: 11