Get rid of saveIter in leaf_data
I can use semantic actions to fail the parser (and let Spirit revert the
iterator) in case the value is not valid. I wanted to this before, and
there was a problem where the attribute wasn't getting properly
assigned, but I got it working now.  Not sure what I did wrong before...

Change-Id: Idb281ccd303d20be59cbef381ff824947b6b11c9