ls: Add test for container inside a list
The first path fragment gets parsed as a data node, but the second
can't. That means the parser has to backtrack and try to parse the path
as a schema path. A later patch had a problem where ParserContext path
wasn't getting properly reset. The actual problem happened in the
describe command, but they take the same type of paths.

Change-Id: Id62884d303c0a97051c78a80144f5f877821af2f