Fix create/delete/set command crash
These commands crashed when they were given an empty path. Right now,
PathParser is allowed to parse empty paths (an empty string or just a
slash). The parsers that check the last node of a path don't expect the
path to be empty. Right now, I chose this easy fix, but PathParser
probably shouldn't be allowed to parse empty paths in the future (and
empty path should be dealt with in a different way).

Issue: https://tree.taiga.io/project/jktjkt-netconf-cli/issue/171
Change-Id: If0a1b97319307f88e83a07914186b6cf8c61b1d2