CI: keep testing the Python bindings
Since https://gerrit.cesnet.cz/c/CzechLight/netconf-cli/+/2550 these
were ignored by the CI because it only matched jobs ending on "-gcc". We
do not have any such jobs since that commit as we're moved to coverage
even before that, and the "f29-gcc" was the only one remaining that
matched. Oops.

Change-Id: Ibe572bca6cbcd02e26dc1f4424694f12d5cac957