Change leaf_data_ to std::variant
Boost Spirit kind of doesn't support std::variant in a weird way. The
problem is that even though leaf_data_enum has and attribute of enum_,
it still gets saved as an std::string. I have no idea why this happens,
but a Github issue explains and a patch from a comment solves the
problem. However, as I have no idea what the patch does, I won't include it.

Change-Id: Ie9ea7b60c6d837b1961afe791db2cdf9da90a573